Added: admin feautures

+++ internal/core/service/user_service.go
@@ -3,6 +3,7 @@ package service
 import (
 	"auth-microservice/internal/core/domain"
 	"auth-microservice/internal/core/ports"
+	"errors"
 )
 
 type UserService struct {
@@ -18,6 +19,9 @@ func NewService(usersRepository ports.UsersRepository, passwordCheckerRepository
 }
 
 func (service *UserService) UserByID(id int64) (*domain.User, error) {
+	if id < 1 {
+		return nil, ErrInvalidId
+	}
 	return service.UsersRepository.UserByID(id)
 }
 
@@ -49,5 +53,36 @@ func (service *UserService) UserByLogin(login string, password string) (*domain.
 }
 
 func (service *UserService) DeleteUser(id int64) error {
+	if id < 1 {
+		return ErrInvalidId
+	}
 	return service.UsersRepository.DeleteUser(id)
+}
+
+func (service *UserService) AllUsers() ([]*domain.User, error) {
+	users, err := service.UsersRepository.GetAllUsers()
+	if err != nil {
+		if errors.Is(err, ports.ErrFailedToLoad) {
+			return []*domain.User{}, ErrInternalServer
+		}
+		return []*domain.User{}, err
+	}
+	return users, nil
+}
+
+func (service *UserService) UpdateUser(userId int64, role domain.Role) error {
+	if userId < 1 {
+		return ErrInvalidId
+	}
+	err := service.UsersRepository.UpdateUser(userId, role)
+	if err != nil {
+		if errors.Is(err, ports.ErrFailedToLoad) {
+			return ErrInternalServer
+		}
+		if errors.Is(err, ports.ErrFailedToUpdate) {
+			return ErrInternalServer
+		}
+		return err
+	}
+	return nil
 }
